refactoring: there was splitted database.service on pices where each api endpoint responsible for itself
