+++ src/interfaces/IArtis.ts
@@ -1,0 +1,5 @@
+interface IArtist {
+  id: string; // uuid v4
+  name: string;
+  grammy: boolean;
+}
