+++ src/interfaces/ITrack.ts
@@ -1,0 +1,7 @@
+export interface ITrack {
+  id: string; // uuid v4
+  name: string;
+  artistId: string | null; // refers to Artist
+  albumId: string | null; // refers to Album
+  duration: number; // integer number
+}
